refactor(api): isolate /test/* endpoints into a partial class, gate off in production
- Move the 6 /test/* handlers + their 9 DTOs out of ApiService.cs into
  ApiService.TestEndpoints.cs / ApiService.TestEndpoints.Models.cs (partial class).
- Add an early-return dispatcher gate: /test/* returns 404 in production
  (Env.IsTest false), dispatches normally under test. Placed after the auth
  check so production behaves identically to an unknown route.
- Add shared WriteNotFoundAsync helper; route the 3 existing default 404 arms
  through it for a byte-identical 404.
- Add optional includeMethod predicate to OpenApiGenerator.Generate; exclude
  /test/* from the spec when not under test, sharing the dispatcher's gate so
  the published contract and routing can't drift.
- Fix tools/openapi-generator reflection caller: pass the new argument
  explicitly (MethodInfo.Invoke ignores C# optional defaults) and read only the
  typed ApiEndpointAttribute (enumerating all attributes resolved the async
  state-machine type across the net10-tool/net6-mod boundary and failed the
  image build).
